fix: conversion

+++ bytecode.cpp
@@ -360,7 +360,7 @@ void bc_disasm(PycRef<PycCode> code, PycModule* mod, int indent)
             } else if (Pyc::IsJumpOffsetArg(opcode)) {
                 fprintf(pyc_output, "%d (to %d)", operand, pos+operand);
             } else if (Pyc::IsCompareArg(opcode)) {
-                if (operand < cmp_strings_len)
+                if (static_cast<size_t>(operand) < cmp_strings_len)
                     fprintf(pyc_output, "%d (%s)", operand, cmp_strings[operand]);
                 else
                     fprintf(pyc_output, "%d (UNKNOWN)", operand);
